Update derive macro tests
Merge them into a single integration test.
Gate the integration test on the `macros` feature.
Use the test functions from utils.
