+++ tests/derives/lib.rs
@@ -1,0 +1,7 @@
+mod deserialize_fromstr;
+mod serialize_display;
+#[path = "../utils.rs"]
+mod utils;
+
+use expect_test::expect;
+use utils::*;
